% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fto_locus_df.R
\docType{data}
\name{fto_locus_df}
\alias{fto_locus_df}
\title{Example genome wide association study results at the FTO locus}
\format{
A data frame with 19119 rows and 9 variables:
\describe{
\item{chromosome}{chromosome}
\item{position}{position on chromosome}
\item{rsid}{rsid number from dbSNP}
\item{effect_allele}{effect allele (or reference allele)}
\item{other_allele}{non-effect allele (or alternate allele)}
\item{eaf}{effect allele frequenct (or minor allele frequency)}
\item{effect}{effect of allele on outcome}
\item{std_err}{standard error of the effect estimate}
\item{p_value}{p-value of effect estimate}
}
}
\usage{
fto_locus_df
}
\description{
A dataset containing genome wide association study stummary results at the FTO locus on chromosome 16
}
\keyword{datasets}
