% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_geneplot.R
\name{gg_geneplot}
\alias{gg_geneplot}
\title{Plot genes located within a genomic region of interest}
\usage{
gg_geneplot(chr, start, end, genome_build = "GRCh38", max_levels = 5)
}
\arguments{
\item{chr}{Integer - chromosome}

\item{start}{Integer - starting position for region of interest}

\item{end}{Integer - ending position for region of interest}

\item{genome_build}{Character - genome build - one of "GRCh37" or "GRCh38"}

\item{max_levels}{Integer - maximum number of levels for gene tracks}
}
\value{
A ggplot object containing a plot of genes within the region of interest
}
\description{
Returns a ggplot containing the genes within a specified genomic region. The function uses database connections to EnsDb.Hsapiens.v75 (hg19/GRCh37) or EnsDb.Hsapiens.v86 (hg38/GRCh38) to identify genes within the specified region, and uses the ggbio package to create the plot.
}
\examples{
\dontrun{
gg_geneplot(1, 170054349 - 1e6, 170054349 + 1e6, "GRCh37")
}

}
