% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggregionalassoc-package.R
\docType{package}
\name{ggregionalassoc-package}
\alias{ggregionalassoc}
\alias{ggregionalassoc-package}
\title{ggregionalassoc: Create Regional Association Plots}
\description{
Description: This package allows the user to obtain linkage disequlibrium data for any contiguous region within the genome using the LocusZoom API \href{https://portaldev.sph.umich.edu/}{https://portaldev.sph.umich.edu/}. By integrating this LD information with results from a genome-wide association study (GWAS) the user can create regional association plots.
}
\author{
\strong{Maintainer}: Michael Levin \email{mglev1n@gmail.com} (\href{https://orcid.org/0000-0002-9937-9932}{ORCID})

}
\keyword{internal}
