% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_extract_locuszoom.R
\name{ld_extract_locuszoom}
\alias{ld_extract_locuszoom}
\title{ld_extract_locuszoom}
\usage{
ld_extract_locuszoom(
  chrom,
  pos,
  ref,
  alt,
  start,
  stop,
  build = "GRCh37",
  population = "ALL",
  metric = "rsquare"
)
}
\arguments{
\item{chrom}{Integer - chromosome of reference variant}

\item{pos}{Integer - position of reference variant}

\item{ref}{Character - reference allele (or effect allele) for reference variant}

\item{alt}{Character - alternate allele (or non-effect allele) for reference variant}

\item{start}{Integer - starting position of range of interest}

\item{stop}{Integer - ending position of range of interest}

\item{build}{Character - one of "GRCh37" or "GRCh38"}

\item{population}{Character - one of "ALL", "AFR", "AMR", "EAS", "EUR", "SAS"}

\item{metric}{Character - one of "r", "rsquare", or "cov", referring to the correlation statistic of interest}
}
\value{
A tibble containing each variant within the supplied range surrounding the variant of interest, with the requested linkage disequilibrium information with respect to the variant of interest
}
\description{
This function allows the user to extract linkage disequilibrium statistics between a variant of interest and surrounding variants within a contiguous genomic region. This function uses the University of Michigan LocusZoom API (\url{https://portaldev.sph.umich.edu/}) to obtain LD information, and allows the user to specify genome-build and ancestry of interest.
}
\examples{
\dontrun{
ld_extract_locuszoom(chrom = 16, pos = 53830055, ref = "C", alt = "G", start = 53830055 - 5e5, stop = 53830055 + 5e5, build = "GRCh37", population = "ALL", metric = "rsquare")
}
}
