% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locusplotr-package.R
\docType{package}
\name{locusplotr-package}
\alias{locusplotr}
\alias{locusplotr-package}
\title{locusplotr: Create Regional Association Plots}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Description: This package allows the user to obtain linkage disequlibrium data for any contiguous region within the genome using the LocusZoom API \href{https://portaldev.sph.umich.edu/}{https://portaldev.sph.umich.edu/}. By integrating this LD information with results from a genome-wide association study (GWAS) the user can create regional association plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mglev1n.github.io/locusplotr}
  \item \url{https://github.com/mglev1n/locusplotr}
}

}
\author{
\strong{Maintainer}: Michael Levin \email{mglev1n@gmail.com} (\href{https://orcid.org/0000-0002-9937-9932}{ORCID})

}
\keyword{internal}
