% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recomb_extract_locuszoom.R
\name{recomb_extract_locuszoom}
\alias{recomb_extract_locuszoom}
\title{Query LocusZoom API for Recombination Data}
\usage{
recomb_extract_locuszoom(chrom, start, end, genome_build = "GRCh37")
}
\arguments{
\item{chrom}{A numeric value specifying the chromosome (e.g., 1, 2, ..., 22, 23 for X, 24 for Y)}

\item{start}{An integer specifying the start position of the region of interest}

\item{end}{An integer specifying the end position of the region of interest}

\item{genome_build}{A character string specifying the genome build (default: "GRCh37")}
}
\value{
A tibble containing the parsed recombination data from the LocusZoom API
}
\description{
This function queries the LocusZoom API to retrieve recombination data
for a specified genomic region and returns the result as a tibble.
}
\examples{
\dontrun{
result <- recomb_locuszoom(chrom = 1, start = 1000, end = 150000)
print(result)

# Using a different genome build
result_grch38 <- recomb_locuszoom(chrom = 1, start = 1000, end = 150000, genome_build = "GRCh38")
}

}
