% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_locusplot.R
\name{gg_locusplot}
\alias{gg_locusplot}
\title{gg_locusplot}
\usage{
gg_locusplot(
  df,
  lead_snps,
  rsid = rsid,
  chromosome = chromosome,
  position = position,
  ref = ref,
  alt = alt,
  p_value = p_value,
  plot_pvalue_threshold = 0.1,
  plot_subsample_prop = 0.1,
  plot_distance = 5e+05,
  genome_build = "GRCh37",
  population = "ALL",
  plot_genes = FALSE,
  plot_title = NULL,
  plot_subtitle = NULL,
  path = NULL
)
}
\arguments{
\item{df}{Dataframe containing columns with rsid, chromosome, position, reference/effect allele, alternate/non-effect allele, and p-value for all variants within the range of interest}

\item{lead_snps}{A character vector containing the lead variant of interest}

\item{rsid}{Rsid column}

\item{chromosome}{Chromosome column}

\item{position}{Position column}

\item{ref}{Reference/effect allele column}

\item{alt}{Alternate/non-effect allele column}

\item{p_value}{P-value column}

\item{plot_pvalue_threshold}{Threshold for plotting p-value on regional association plot (default = 0.1) - reducing the number of points decreases file size and improves performance}

\item{plot_subsample_prop}{Proportion of points above p-value threshold to plot (default = 0.1) - reducing the number of points decreases file size and improves performance}

\item{plot_distance}{Integer corresponding to the size of the locus that should be plotted}

\item{genome_build}{Character - one of "GRCh37" or "GRCh38"}

\item{population}{Character - one of "ALL", "AFR", "AMR", "EAS", "EUR", "SAS" referring to the reference population of interest for obtaining linkage disequilibrium information (default = "ALL")}

\item{plot_genes}{Logical - Include a plot of genes/transcripts within the region of interest beneath the regional association plot (default = FALSE)}

\item{plot_title}{A character string corresponding to plot title (default = NULL)}

\item{plot_subtitle}{A character string corresponding to plot subtitle (default = NULL)}

\item{path}{Character string (default = NULL) - if a path is supplied a .pdf of the plot will be saved}
}
\value{
A ggplot object
}
\description{
Returns a ggplot object containing a regional association plot (-log10(p-value) as a function of chromosomal position, with variants colored by linkage disequilibrium to reference variant).
This function allows the user to integrate genome wide association study (GWAS) summary statistics for a locus of interest with linkage disequilibrium information (obtained using the University of Michigan LocusZoom API \url{https://portaldev.sph.umich.edu/}) for that locus to create a regional association plot.
}
\examples{
\dontrun{
library(locusplotr)
library(tidyverse)
test <- arrow::read_parquet("../Varicose-Veins-GWAMA/Data/HyPrColoc/vv_risk_loci_transancestry_20210324_FUMA_20210928.parquet") \%>\%
  filter(lead_rsid == "rs6025")
test \%>\%
  gg_regional_assoc(lead_snps = "rs6025", , rsid = rsid, chromosome = CHR, position = POSITION, ref = Allele1, alt = Allele2, p_value = `P-value`, plot_genes = TRUE, plot_title = NULL, plot_subtitle = NULL, plot_distance = 1e6, path = NULL)
}
}
