% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_locusplot.R
\name{gg_locusplot}
\alias{gg_locusplot}
\title{Create a regional association plot}
\usage{
gg_locusplot(
  df,
  lead_snp = NULL,
  rsid = rsid,
  chrom = chrom,
  pos = pos,
  ref = ref,
  alt = alt,
  p_value = p_value,
  plot_pvalue_threshold = 0.1,
  plot_subsample_prop = 0.1,
  plot_distance = 5e+05,
  genome_build = "GRCh37",
  population = "ALL",
  plot_genes = FALSE,
  plot_title = NULL,
  plot_subtitle = NULL,
  path = NULL
)
}
\arguments{
\item{df}{Dataframe containing columns with rsid, chromosome, position, reference/effect allele, alternate/non-effect allele, and p-value for all variants within the range of interest}

\item{lead_snp}{A character vector containing a lead variant of interest. When NULL (default), the variant with the lowest p-value will be selected as the lead variant.}

\item{rsid}{Rsid column}

\item{chrom}{Chromosome column}

\item{pos}{Position column}

\item{ref}{Reference/effect allele column}

\item{alt}{Alternate/non-effect allele column}

\item{p_value}{P-value column}

\item{plot_pvalue_threshold}{Threshold for plotting p-value on regional association plot (default = 0.1) - reducing the number of points decreases file size and improves performance}

\item{plot_subsample_prop}{Proportion of points above p-value threshold to plot (default = 0.1; range = 0-1) - reducing the number of points decreases file size and improves performance}

\item{plot_distance}{Integer corresponding to the size of the locus that should be plotted}

\item{genome_build}{Character - one of "GRCh37" or "GRCh38"}

\item{population}{Character - one of "ALL", "AFR", "AMR", "EAS", "EUR", "SAS" referring to the reference population of interest for obtaining linkage disequilibrium information (default = "ALL")}

\item{plot_genes}{Logical - Include a plot of genes/transcripts within the region of interest beneath the regional association plot (default = FALSE)}

\item{plot_title}{A character string corresponding to plot title (default = NULL)}

\item{plot_subtitle}{A character string corresponding to plot subtitle (default = NULL)}

\item{path}{Character string (default = NULL) - if a path is supplied a .pdf of the plot will be saved}
}
\value{
A ggplot object containing a regional association plot for the locus of interest
}
\description{
Returns a ggplot object containing a regional association plot (-log10(p-value) as a function of chromosomal position, with variants colored by linkage disequilibrium to reference variant).
This function allows the user to integrate genome wide association study (GWAS) summary statistics for a locus of interest with linkage disequilibrium information (obtained using the University of Michigan LocusZoom API \url{https://portaldev.sph.umich.edu/}) for that locus to create a regional association plot.
}
\examples{
\dontrun{
# Basic regional association plot
gg_locusplot(df = fto_locus_df, lead_snp = "rs62033413", rsid = rsid, chrom = chromosome, pos = position, ref = effect_allele, alt = other_allele, p_value = p_value)

# Use "plot_genes = TRUE" to add a plot of genes within the region beneath the regional association plot
gg_locusplot(df = fto_locus_df, lead_snp = "rs62033413", rsid = rsid, chrom = chromosome, pos = position, ref = effect_allele, alt = other_allele, p_value = p_value, plot_genes = TRUE)
}

}
